requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.316
langchain-core==0.3.12
langchain-openai==0.2.3
langsmith==0.1.136
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.1
nibabel==5.3.0
nipype==1.8.6
numexpr==2.10.1
numpy==1.26.4
onnxruntime==1.19.2
openai==0.28.1
orjson==3.10.9
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
posthog==3.7.0
propcache==0.2.0
protobuf==5.28.2
prov==2.0.1
pulsar-client==3.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
PyMuPDF==1.24.11
pyparsing==3.1.4
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
scipy==1.14.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
traits==6.3.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
yarl==1.15.0
